Add links, refactor/fix import
